requirements.txt
Flask==0.10.1
gunicorn==18.0
honcho==0.5.0
Pillow==2.2.2
zbar==0.10
